adde update settings method for filing + seperate dataset_settings methods
