+++ development/develop_script.py
@@ -41,41 +41,5 @@ dataset.update_settings(input_data_file=testdatafile,
                         output_folder='/home/thoverga/Documents'
                         )
 
-obstype = 'temp'
 
-
-
-print('before : ', dataset.settings.qc['qc_check_settings']["gross_value"][obstype]['min_value'])
-
-dataset.update_qc_settings(obstype=obstype, gross_value_min_value=-100)
-
-
-print('after: ', dataset.settings.qc['qc_check_settings']["gross_value"][obstype]['min_value'])
-
-
-
-#%% Datset2
-
-
-dataset2 = metobs_toolkit.Dataset()
-dataset2.update_settings(input_data_file=testdatafile,
-                        output_folder='/home/thoverga/Documents'
-                        )
-
-print('before2 : ', dataset2.settings.qc['qc_check_settings']["gross_value"][obstype]['min_value'])
-
-dataset2.update_qc_settings(obstype=obstype, gross_value_min_value=-100)
-
-
-print('after2: ', dataset2.settings.qc['qc_check_settings']["gross_value"][obstype]['min_value'])
-
-
-
-
-#%%
-
-from metobs_toolkit.settings import Settings
-
-testset = Settings()
-print('after settings creation: ', testset.qc['qc_check_settings']["gross_value"][obstype]['min_value'])
