+++ metobs_toolkit/dataset.py
@@ -116,207 +116,7 @@ class Dataset:
         self.settings = copy.deepcopy(Settings())
 
 
-    def update_settings(
-        self,
-        output_folder=None,
-        input_data_file=None,
-        input_metadata_file=None,
-        data_template_file=None,
-        metadata_template_file=None,
-    ):
-        """
-        Update the most common input-output (IO) settings.
-        (This should be applied before importing the observations.)
 
-        When an update value is None, the specific setting will not be updated.
-
-        Parameters
-        ----------
-        output_folder : string, optional
-            A directory to store the output to. The default is None.
-        input_data_file : string, optional
-            Path to the input data file with observations. The default is None.
-        input_metadata_file : string, optional
-            Path to the input metadata file. The default is None.
-        data_template_file : string, optional
-            Path to the mapper-template csv file to be used on the observations.. The default is None.
-        metadata_template_file : string, optional
-            Path to the mapper-template csv file to be used on the metadata.. The default is None.
-
-        Returns
-        -------
-        None.
-
-        """
-
-        self.settings.update_IO(
-            output_folder=output_folder,
-            input_data_file=input_data_file,
-            input_metadata_file=input_metadata_file,
-            data_template_file=data_template_file,
-            metadata_template_file=metadata_template_file,
-        )
-
-
-    def update_timezone(self, timezonestr):
-        """
-        Change the timezone of the input data. By default the Brussels timezone is assumed.
-        A valid timezonestring is an element of the pytz.all_timezones.
-
-        Parameters
-        ----------
-        timezonestr : string
-            Timezone string of the input observations. Element of pytz.all_timezones.
-
-        Returns
-        -------
-        None.
-
-        """
-
-        self.settings.update_timezone(timezonestr)
-
-    def update_default_name(self, default_name):
-        """
-        Update the default name (the name of the station). This name will be
-        used when no names are found in the observational dataset.
-
-        (All observations are assumed to come from one station.)
-
-        Parameters
-        ----------
-        default_name : string
-            Default name to use when no names are present in the data.
-
-        Returns
-        -------
-        None.
-
-        """
-
-        self.settings.app["default_name"] = str(default_name)
-
-    def update_qc_settings(self, obstype='temp', dupl_timestamp_keep=None, persis_time_win_to_check=None, persis_min_num_obs=None,
-                            rep_max_valid_repetitions=None, gross_value_min_value=None, gross_value_max_value=None,
-                            win_var_max_increase_per_sec=None, win_var_max_decrease_per_sec=None, win_var_time_win_to_check=None,
-                            win_var_min_num_obs=None, step_max_increase_per_sec=None, step_max_decrease_per_sec=None):
-
-        """
-        Update the QC settings for the specified observation type.
-
-        If a argument value is None, the default settings will not be updated.
-        Parameters
-        ----------
-        dupl_timestamp_keep : bool
-            Setting that determines to keep, or remove duplicated timestamps.
-        persis_time_win_to_check : Timedelta or str
-            Time window for persistance check
-        persis_min_num_obs : int (> 0)
-            Minimal window members for persistance check
-        rep_max_valid_repetitions : int (> 0)
-            Maximal valid repetitions for repetitions check
-        gross_value_min_value : numeric
-            Minimal value for gross value check
-        gross_value_max_value : numeric
-            Maximal value for gross value check
-        win_var_max_increase_per_sec : numeric (> 0)
-            Maximal increase per second for window variation check
-        win_var_max_decrease_per_sec : numeric (> 0)
-            Maximal decrease per second for window variation check
-        win_var_time_win_to_check : Timedelta or str
-            Time window for window variation check
-        win_var_min_num_obs : int (> 0)
-            Minimal window members for window variation check
-        step_max_increase_per_sec : numeric
-            Maximal increase per second for step check
-        step_max_decrease_per_sec : numeric (< 0)
-            Maximal decrease per second for step check
-
-        Returns
-        -------
-        None.
-
-        """
-        def is_timedelta(timedeltastr):
-            try:
-                pd.to_timedelta(timedeltastr)
-                return True
-            except:
-                return False
-
-        assert obstype in observation_types, f'{obstype} is not a known observation type'
-
-        # Gross value check
-        if not gross_value_max_value is None:
-            logger.info(f'Maximal value for gross value check updated: \
-        {self.settings.qc["qc_check_settings"]["gross_value"][obstype]["max_value"]} --> {float(gross_value_max_value)}')
-            self.settings.qc['qc_check_settings']["gross_value"][obstype]['max_value'] = float(gross_value_max_value)
-
-        if not gross_value_min_value is None:
-            logger.info(f'Manimal value for gross value check updated: \
-        {self.settings.qc["qc_check_settings"]["gross_value"][obstype]["min_value"]} --> {float(gross_value_min_value)}')
-            self.settings.qc['qc_check_settings']["gross_value"][obstype]['min_value'] = float(gross_value_min_value)
-
-        # Duplicate check
-        if not dupl_timestamp_keep is None:
-            logger.info(f'Setting to keep (True) are remove (False) duplicate timestamps updated: \
-        {self.settings.qc["qc_check_settings"]["duplicated_timestamp"]["keep"]} -->  {bool(dupl_timestamp_keep)}')
-            self.settings.qc['qc_check_settings']["duplicated_timestamp"]['keep'] = bool(dupl_timestamp_keep)
-
-        # Persistance check
-        if not persis_time_win_to_check is None:
-            if is_timedelta(str(persis_time_win_to_check)):
-                logger.info(f'Time window size for persistance check updated:\
-                {self.settings.qc["qc_check_settings"]["persistance"][obstype]["time_window_to_check"]}--> {str(persis_time_win_to_check)}')
-                self.settings.qc['qc_check_settings']["persistance"][obstype]['time_window_to_check'] = str(persis_time_win_to_check)
-            else:
-                logger.warning(f' {str(persis_time_win_to_check)} is not a valid timedelta string. No update on this setting.')
-
-        if not persis_min_num_obs is None:
-            logger.info(f'Minimal window members for persistance check updated:\
-            {self.settings.qc["qc_check_settings"]["persistance"][obstype]["min_num_obs"]} --> {abs(int(persis_min_num_obs))}')
-            self.settings.qc['qc_check_settings']["persistance"][obstype]['min_num_obs'] = abs(int(persis_min_num_obs))
-
-        # Repetitions check
-        if not rep_max_valid_repetitions is None:
-            logger.info(f'Maximal valid repetitions for repetitions check updated: \
-                        {self.settings.qc["qc_check_settings"]["repetitions"][obstype]["max_valid_repetitions"]} --> {abs(int(rep_max_valid_repetitions))}')
-            self.settings.qc['qc_check_settings']["repetitions"][obstype]['max_valid_repetitions'] = abs(int(rep_max_valid_repetitions))
-
-        # Window variation check
-        if not win_var_max_increase_per_sec is None:
-            logger.info(f'Maximal increase per second for window variation check updated:\
-                        {self.settings.qc["qc_check_settings"]["window_variation"][obstype]["max_increase_per_second"]} --> {abs(float(win_var_max_increase_per_sec))}')
-            self.settings.qc['qc_check_settings']["window_variation"][obstype]['max_increase_per_second'] = abs(float(win_var_max_increase_per_sec))
-
-        if not win_var_max_decrease_per_sec is None:
-            logger.info(f'Maximal decrease per second for window variation check updated:\
-                        {self.settings.qc["qc_check_settings"]["window_variation"][obstype]["max_decrease_per_second"] } --> {abs(float(win_var_max_decrease_per_sec))}')
-            self.settings.qc['qc_check_settings']["window_variation"][obstype]['max_decrease_per_second'] = abs(float(win_var_max_decrease_per_sec))
-
-        if not win_var_time_win_to_check is None:
-            if is_timedelta(str(win_var_time_win_to_check)):
-                logger.info(f'Time window for window variation check updated:\
-                            {self.settings.qc["qc_check_settings"]["window_variation"][obstype]["time_window_to_check"]} --> {str(win_var_time_win_to_check)}')
-                self.settings.qc['qc_check_settings']["window_variation"][obstype]['time_window_to_check'] = str(win_var_time_win_to_check)
-            else:
-                logger.warning(f' {str(persis_time_win_to_check)} is not a valid timedelta string. No update on this setting.')
-
-        if not win_var_min_num_obs is None:
-            logger.info(f'Minimal window members for window variation check updated:\
-                         {self.settings.qc["qc_check_settings"]["window_variation"][obstype]["min_window_members"]}--> {abs(int(win_var_min_num_obs))}')
-            self.settings.qc['qc_check_settings']["window_variation"][obstype]['min_window_members'] = abs(int(win_var_min_num_obs))
-
-        # Step check
-        if not step_max_increase_per_sec is None:
-            logger.info(f'Maximal increase per second for step check updated:\
-                        {self.settings.qc["qc_check_settings"]["step"][obstype]["max_increase_per_second"]}--> {abs(float(step_max_increase_per_sec))}')
-            self.settings.qc['qc_check_settings']["step"][obstype]['max_increase_per_second'] = abs(float(step_max_increase_per_sec))
-
-        if not step_max_decrease_per_sec is None:
-            logger.info(f'Maximal decrease per second for step check updated:\
-                       {self.settings.qc["qc_check_settings"]["step"][obstype]["max_decrease_per_second"]} --> {-1.0 * abs(float(step_max_decrease_per_sec))}')
-            self.settings.qc['qc_check_settings']["step"][obstype]['max_decrease_per_second'] = -1.0 * abs(float(step_max_decrease_per_sec))
 
 
     def show_settings(self):
@@ -1502,10 +1302,8 @@ class Dataset:
 
         """
 
-        # Start from input_df !!!!!
-        #   reset self.df, and dataset resolution in metadf
-        #   reset self.outliersdf, gaps, missing
-        #
+
+
         df = self.input_df
 
         self.df = init_multiindexdf()
